+++ CharacterDatabaseAPI/src/Models/ConverterUtil.cs
