+++ CharacterDatabaseAPI/src/Properties/launchSettings.json
