--- Program.cs
@@ -1,43 +1,0 @@
-using CharacterDatabaseAPI.Services;
-using Amazon.Extensions.NETCore.Setup;
-using CharacterDatabaseAPI.WebScraper;
-
-var builder = WebApplication.CreateBuilder(args);
-
-// Get the AWS profile information from configuration providers
-AWSOptions awsOptions = builder.Configuration.GetAWSOptions();
-
-// Configure AWS service clients to use these credentials
-builder.Services.AddDefaultAWSOptions(awsOptions);
-
-builder.Services.AddScoped<ICharacterService, CharacterService>();
-
-// Add services to the container.
-
-builder.Services.AddControllers().AddJsonOptions(
-        options => options.JsonSerializerOptions.PropertyNamingPolicy = null);
-
-builder.Configuration.AddEnvironmentVariables();
-
-
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
-{
-    app.UseDeveloperExceptionPage();
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseHttpsRedirection();
-
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.Run();
